backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
gseapy>=1.1.12
pandas>=2.2.3
numpy>=2.0.0
scipy>=1.14.0
matplotlib>=3.9.0
seaborn>=0.13.2
plotly>=5.24.0
kaleido==0.2.1
python-docx==1.1.2
jinja2==3.1.4
openpyxl>=3.1.5
python-multipart==0.0.12
pydantic-settings==2.4.0
scikit-learn>=1.5.2
